@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
-__version__ = "1.2.3"
+__version__ = "1.3.0"
 
 
 def version():
